+++ tests/test_acceptance.py
@@ -18,10 +18,8 @@ from pathlib import Path
 from typing import Dict
 
 import pytest
-from cloudai import NcclTest, Parser, Test, UCCTest
-from cloudai.__main__ import handle_dry_run_and_run, identify_unique_test_templates, setup_logging
-from cloudai.systems import StandaloneSystem
-from cloudai.test_definitions.nccl import NCCLCmdArgs, NCCLTestDefinition
+from cloudai import Parser
+from cloudai.__main__ import handle_dry_run_and_run, setup_logging
 
 SLURM_TEST_SCENARIOS = [
     {"path": Path("conf/common/test_scenario/sleep.toml"), "expected_dirs_number": 4, "log_file": "sleep_debug.log"},
@@ -64,55 +62,3 @@ def test_slurm(tmp_path: Path, scenario: Dict):
         assert "Tests." in td.name, "Invalid test directory name"
 
     assert log_file_path.exists(), f"Log file {log_file_path} was not created"
-
-
-class TestIdentifyUniqueTestTemplates:
-    @pytest.fixture
-    def system(self, tmp_path: Path) -> StandaloneSystem:
-        return StandaloneSystem(name="system", install_path=tmp_path, output_path=tmp_path)
-
-    @pytest.fixture
-    def test_def(self) -> NCCLTestDefinition:
-        return NCCLTestDefinition(name="nccl", description="", test_template_name="ttname", cmd_args=NCCLCmdArgs())
-
-    def test_single_input(self, system: StandaloneSystem, test_def: NCCLTestDefinition):
-        templ = NcclTest(system, "template_name")
-        test = Test(test_definition=test_def, test_template=templ)
-
-        res = identify_unique_test_templates([test])
-
-        assert len(res) == 1
-        assert res[0] == templ
-
-    def test_two_templates_with_different_names(self, system: StandaloneSystem, test_def: NCCLTestDefinition):
-        templ1 = NcclTest(system, "template_name1")
-        templ2 = NcclTest(system, "template_name2")
-        test1 = Test(test_definition=test_def, test_template=templ1)
-        test2 = Test(test_definition=test_def, test_template=templ2)
-
-        res = identify_unique_test_templates([test1, test2])
-
-        assert len(res) == 1
-        assert res[0] == templ1
-
-    def test_two_templates_with_same_name(self, system: StandaloneSystem, test_def: NCCLTestDefinition):
-        templ = NcclTest(system, "template_name")
-        test1 = Test(test_definition=test_def, test_template=templ)
-        test2 = Test(test_definition=test_def, test_template=templ)
-
-        res = identify_unique_test_templates([test1, test2])
-
-        assert len(res) == 1
-        assert res[0] == templ
-
-    def test_two_different_templates_with_same_name(self, system: StandaloneSystem, test_def: NCCLTestDefinition):
-        templ1 = NcclTest(system, "template_name")
-        templ2 = UCCTest(system, "template_name")
-        test1 = Test(test_definition=test_def, test_template=templ1)
-        test2 = Test(test_definition=test_def, test_template=templ2)
-
-        res = identify_unique_test_templates([test1, test2])
-
-        assert len(res) == 2
-        assert templ1 in res
-        assert templ2 in res
